Avoid repeatedly installing Pytools
pip install for the Pytools is executed whenever make is typed in the
top-level directory, even if no changes have been made to the Python
source files. The makefile now creates a timestamp after Pytools is
installed, and reinstalls it only if the source files become newer
than the timestamp.